requirements.txt
asgiref==3.3.1
attrs==20.3.0
Automat==20.2.0
cffi==1.14.5
constantly==15.1.0
cryptography==3.4.6
cssselect==1.1.0
Django==3.0.5
django-countries==7.1
django-taggit==1.3.0
django-widget-tweaks==1.4.8
djongo==1.3.4
hyperlink==21.0.0
idna==3.1
incremental==21.3.0
itemadapter==0.2.0
itemloaders==1.0.4
jmespath==0.10.0
joblib==1.0.1
lxml==4.6.2
numpy==1.20.1
pandas==1.2.3
parsel==1.6.0
Pillow==8.0.1
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pymongo==3.11.3
pyOpenSSL==20.0.1
python-dateutil==2.8.1
pytz==2021.1
queuelib==1.5.0
scikit-learn==0.24.1
scipy==1.6.1
Scrapy==2.4.1
scrapy-djangoitem==1.1.1
service-identity==18.1.0
six==1.15.0
sklearn==0.0
sqlparse==0.2.4
threadpoolctl==2.1.0
Twisted==21.2.0
twisted-iocpsupport==1.0.1
w3lib==1.22.0
zope.interface==5.2.0
